+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [STVP](https://mia.kiwi/projects/stvp).
 
+## [25.2.3] - 2025-12-05
+
+### Added
+
+- Chips component
+
+### Changed
+
+- Made icon-only buttons have smaller padding
+
+
+
 ## [25.2.2] - 2025-11-29
 
 ### Added